01_models_prompts_and_parsers/pyproject.toml
[project]
name = "01-models-prompts-and-parsers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "langchain-community>=0.3.15",
    "langchain-google-genai>=2.0.9",
    "python-dotenv>=1.0.1",
]
